backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
certifi
numpy
sentence-transformers
transformers
torch
pinecone-client
groq
python-dotenv
python-multipart
pydantic
reportlab
